number of packets changed from float to int

+++ Shared/ClientData.cs
@@ -11,6 +11,6 @@ namespace Shared
         public float X { get; set; }
         public float Y { get; set; }
         public float Z { get; set; }
-        public float NumberOfPacket { get; set; }
+        public int NumberOfPacket { get; set; }
     }
 }
